+++ cmake/DLAF_AddTest.cmake
@@ -1,5 +1,5 @@
 #
-# NS3C
+# Distributed Linear Algebra with Future (DLAF)
 #
 # Copyright (c) 2018-2019, ETH Zurich
 # All rights reserved.
@@ -8,7 +8,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
-# NS3C_addTest(test_target_name
+# DLAF_addTest(test_target_name
 #   SOURCES <source1> [<source2> ...]
 #   [COMPILE_DEFINITIONS <arguments for target_compile_definitions>]
 #   [INCLUDE_DIRS <arguments for target_include_directories>]
@@ -26,7 +26,7 @@
 #
 # e.g.
 #
-# NS3C_addTest(example_test
+# DLAF_addTest(example_test
 #   SOURCE main.cpp testfixture.cpp
 #   LIBRARIES
 #     PRIVATE
@@ -34,59 +34,59 @@
 #       include/
 # )
 
-function(NS3C_addTest test_target_name)
+function(DLAF_addTest test_target_name)
   set(options "USE_GTEST_MAIN")
   set(oneValueArgs MPIRANKS)
   set(multiValueArgs SOURCES COMPILE_DEFINITIONS INCLUDE_DIRS LIBRARIES ARGUMENTS)
-  cmake_parse_arguments(NS3C_AT "${options}" "${oneValueArgs}" "${multiValueArgs}" ${ARGN})
+  cmake_parse_arguments(DLAF_AT "${options}" "${oneValueArgs}" "${multiValueArgs}" ${ARGN})
 
   ### Checks
-  if (NS3C_AT_UNPARSED_ARGUMENTS)
-    message(FATAL_ERROR "Unknown arguments ${NS3C_AT_UNPARSED_ARGUMENTS}")
+  if (DLAF_AT_UNPARSED_ARGUMENTS)
+    message(FATAL_ERROR "Unknown arguments ${DLAF_AT_UNPARSED_ARGUMENTS}")
   endif()
 
-  if (NOT NS3C_AT_SOURCES)
+  if (NOT DLAF_AT_SOURCES)
     message(FATAL_ERROR "No sources specified for this test")
   endif()
 
   ### Test executable target
-  add_executable(${test_target_name} ${NS3C_AT_SOURCES})
+  add_executable(${test_target_name} ${DLAF_AT_SOURCES})
 
-  if (NS3C_AT_COMPILE_DEFINITIONS)
+  if (DLAF_AT_COMPILE_DEFINITIONS)
     target_compile_definitions(${test_target_name}
       PRIVATE
-        ${NS3C_AT_COMPILE_DEFINITIONS}
+        ${DLAF_AT_COMPILE_DEFINITIONS}
     )
   endif()
 
-  if (NS3C_AT_INCLUDE_DIRS)
+  if (DLAF_AT_INCLUDE_DIRS)
     target_include_directories(${test_target_name}
-      ${NS3C_AT_INCLUDE_DIRS}
+      ${DLAF_AT_INCLUDE_DIRS}
     )
   endif()
 
-  target_link_libraries(${test_target_name} PRIVATE NS3C)
-  target_link_libraries(${test_target_name} PRIVATE NS3C_test)
-  if (NS3C_AT_USE_GTEST_MAIN)
+  target_link_libraries(${test_target_name} PRIVATE DLAF)
+  target_link_libraries(${test_target_name} PRIVATE DLAF_test)
+  if (DLAF_AT_USE_GTEST_MAIN)
     target_link_libraries(${test_target_name} PRIVATE gtest_main)
   else()
     target_link_libraries(${test_target_name} PRIVATE gtest)
   endif()
 
-  if (NS3C_AT_LIBRARIES)
+  if (DLAF_AT_LIBRARIES)
     target_link_libraries(${test_target_name}
-      ${NS3C_AT_LIBRARIES}
+      ${DLAF_AT_LIBRARIES}
     )
   endif()
 
   ### Test target
   # ----- MPI-based test
-  if(DEFINED NS3C_AT_MPIRANKS)
-    if (NOT NS3C_AT_MPIRANKS GREATER 0)
-      message(FATAL_ERROR "Wrong MPIRANKS number ${NS3C_AT_MPIRANKS}")
+  if(DEFINED DLAF_AT_MPIRANKS)
+    if (NOT DLAF_AT_MPIRANKS GREATER 0)
+      message(FATAL_ERROR "Wrong MPIRANKS number ${DLAF_AT_MPIRANKS}")
     endif()
 
-    if (NS3C_AT_MPIRANKS GREATER MPIEXEC_MAX_NUMPROCS)
+    if (DLAF_AT_MPIRANKS GREATER MPIEXEC_MAX_NUMPROCS)
       message(FATAL_ERROR "Impossible to have more than ${MPIEXEC_MAX_NUMPROCS}")
     endif()
 
@@ -97,13 +97,13 @@ function(NS3C_addTest test_target_name)
     add_test(
       NAME ${test_target_name}
       COMMAND
-        ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} ${NS3C_AT_MPIRANKS}
-        ${MPIEXEC_PREFLAGS} $<TARGET_FILE:${test_target_name}> ${MPIEXEC_POSTFLAGS} ${NS3C_AT_ARGUMENTS})
+        ${MPIEXEC_EXECUTABLE} ${MPIEXEC_NUMPROC_FLAG} ${DLAF_AT_MPIRANKS}
+        ${MPIEXEC_PREFLAGS} $<TARGET_FILE:${test_target_name}> ${MPIEXEC_POSTFLAGS} ${DLAF_AT_ARGUMENTS})
   # ----- Classic test
   else()
     add_test(
       NAME ${test_target_name}
-      COMMAND ${test_target_name} ${NS3C_AT_ARGUMENTS}
+      COMMAND ${test_target_name} ${DLAF_AT_ARGUMENTS}
     )
   endif()
 endfunction()
