+++ include/dlaf/types.h
@@ -1,5 +1,5 @@
 //
-// NS3C
+// Distributed Linear Algebra with Future (DLAF)
 //
 // Copyright (c) 2018-2019, ETH Zurich
 // All rights reserved.
@@ -10,7 +10,7 @@
 
 #pragma once
 
-namespace ns3c {
+namespace dlaf {
 
 using SizeType = int;
