+++ include/dlaf/util_math.h
@@ -1,5 +1,5 @@
 //
-// NS3C
+// Distributed Linear Algebra with Future (DLAF)
 //
 // Copyright (c) 2018-2019, ETH Zurich
 // All rights reserved.
@@ -12,7 +12,7 @@
 
 /// @file
 
-namespace ns3c {
+namespace dlaf {
 namespace util {
 
 /// @brief Returns @c ceiling(@p num / @p den) for integer types.
