+++ test/CMakeLists.txt
@@ -1,5 +1,5 @@
 #
-# NS3C
+# Distributed Linear Algebra with Future (DLAF)
 #
 # Copyright (c) 2018-2019, ETH Zurich
 # All rights reserved.
@@ -8,8 +8,8 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
-add_library(NS3C_test INTERFACE)
-target_include_directories(NS3C_test 
+add_library(DLAF_test INTERFACE)
+target_include_directories(DLAF_test 
     INTERFACE 
     $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include/>
 )
