+++ test/include/dlaf_test/util_types.h
@@ -11,7 +11,7 @@
 #include <complex>
 #include "gtest/gtest.h"
 
-namespace ns3c_test {
+namespace dlaf_test {
 
 using ElementTypes =
     ::testing::Types<int, long long, float, double, std::complex<float>, std::complex<double>>;
