+++ test/unit/memory/CMakeLists.txt
@@ -1,5 +1,5 @@
 #
-# NS3C
+# Distributed Linear Algebra with Future (DLAF)
 #
 # Copyright (c) 2018-2019, ETH Zurich
 # All rights reserved.
@@ -8,5 +8,5 @@
 # SPDX-License-Identifier: BSD-3-Clause
 #
 
-NS3C_addTest(test_memory_chunk SOURCES test_memory_chunk.cpp USE_GTEST_MAIN)
-NS3C_addTest(test_memory_view SOURCES test_memory_view.cpp USE_GTEST_MAIN)
+DLAF_addTest(test_memory_chunk SOURCES test_memory_chunk.cpp USE_GTEST_MAIN)
+DLAF_addTest(test_memory_view SOURCES test_memory_view.cpp USE_GTEST_MAIN)
